Refactored enum parsing to use Optional<>, added new guild level permission exception that extend ParseException.
